NFC: split timepoint fence/semaphore (pool) into separate files
